Move example-only asserts out of lib
The only asserts that should exist inside the distribution of jsclient
should be ones that are actually used by jsclient.
